fastapi/requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
bcrypt==4.1.2
blinker==1.7.0
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
ecdsa==0.18.0
exceptiongroup==1.2.0
fastapi==0.110.0
Flask==3.0.2
Flask-Bcrypt==1.0.1
greenlet==3.0.3
h11==0.14.0
idna==3.6
importlib_metadata==7.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.2
MarkupSafe==2.1.5
mysql==0.0.3
mysql-connector-python==8.3.0
mysqlclient==2.2.4
passlib==1.7.4
psycopg2==2.9.9
pyasn1==0.6.0
pycparser==2.22
pydantic==2.6.4
pydantic-settings==2.2.1
pydantic_core==2.16.3
python-dotenv==1.0.1
python-jose==3.3.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.36.3
typing_extensions==4.10.0
uuid==1.30
uvicorn==0.29.0
Werkzeug==3.0.1
zipp==3.18.1
